requirements.txt
flask==2.3.3
streamlit==1.38.0
twilio==9.2.0
openai==1.30.0
playht==0.2.0
spacy==3.7.2
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1